If the treasure is an enchanted book, enchant it.
Added a bit of logic to check and see if the drop is an enchanted book, and if it is, force at least one enchantment to be applied.  As a warning, I'm not a programmer, so I could have incorrectly used the continue statement here.  This requires more experienced eyes and some testing, I'm sure.